+++ src/main/java/de/fh_zwickau/informatik/sensor/IZWayApi.java
@@ -73,6 +73,12 @@ public interface IZWayApi {
 
     // no asynchronous version for login call
 
+    // Current profile
+    /**
+     * @return ZAutomation profile
+     */
+    Profile getCurrentProfile();
+
     // Namespaces
     /**
      * @return ZAutomation namespace list