+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2017-01-16
+### Added
+- Profile API
+- Add function for getting current profile
+
+### Fixed
+- Handling of different notification types
+
 ## [0.1.9] - 2017-01-09
 ### Added
 - User image field to location